requirements.txt
asyncpg>=0.29.0
uvicorn>=0.30.0
openai-agents==0.0.19
python-dotenv>=1.0.0
fastapi==0.115.12